+++ src/main/java/de/mossgrabers/framework/daw/data/IChannel.java
@@ -1,5 +1,5 @@
 // Written by Jürgen Moßgraber - mossgrabers.de
-// (c) 2017-2018
+// (c) 2017-2019
 // Licensed under LGPLv3 - http://www.gnu.org/licenses/lgpl-3.0.txt
 
 package de.mossgrabers.framework.daw.data;
@@ -69,7 +69,7 @@ public interface IChannel extends IItem
      *
      * @param value The new value
      */
-    void setVolume (double value);
+    void setVolume (int value);
 
 
     /**
@@ -140,7 +140,7 @@ public interface IChannel extends IItem
      *
      * @param value The new value
      */
-    void setPan (double value);
+    void setPan (int value);
 
 
     /**
