+++ src/main/java/de/mossgrabers/framework/daw/data/IParameter.java
@@ -1,5 +1,5 @@
 // Written by Jürgen Moßgraber - mossgrabers.de
-// (c) 2017-2018
+// (c) 2017-2019
 // Licensed under LGPLv3 - http://www.gnu.org/licenses/lgpl-3.0.txt
 
 package de.mossgrabers.framework.daw.data;
@@ -49,7 +49,7 @@ public interface IParameter extends IItem
      *
      * @param value The new value to set
      */
-    void setValue (double value);
+    void setValue (int value);
 
 
     /**
